+++ cookbooks/apple_ilife/metadata.rb
@@ -4,5 +4,5 @@ maintainer_email 'software@wesleyan.edu'
 license          'All rights reserved'
 description      'Installs Apple iLife suite'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.2.8'
+version          '0.3.0'
 depends          'dmg'